rename Duke class to AVA Add simple text repetition feature in chatbot.
