--- src/main/java/Duke.java
@@ -1,24 +1,0 @@
-public class Duke {
-    public static void main(String[] args) {
-
-        System.out.println("""
-                ____________________________________________________________\s
-                Hello ! I'm AVA
-                What can I do for you ?
-                ____________________________________________________________
-                Bye.Hope to see you again soon !
-                ____________________________________________________________""");
-    }
-}
-
-
-/*
-custom hello message after tag 0
-Hi,  I'm AVA (Artificial Virtual Assistant).
-I am a virtual personal assistant created by Nikhil.
-I am currently an infant and can't do much 🙁 but
-don't worry I should soon be very capable 😉
-
-Byee
-
- */
